irq: add IRQ_REPLACE_ISR functionality
These changes introduce subsititution process for IRQ handlers
registered via the IRQ_CONNECT procedure.
The macro can be used when the SoC has a custom IRQ handling method
(e.g. due to an additional interrupt controller) but uses
drivers that are already implemented without having that in mind.
It enables to wrap a default handler function of a driver
into a custom method which will be called whenever the interrupt occurs.
The wrapper function must then explicitly call the original ISR.

Signed-off-by: Jan Kowalewski <jkowalewski@antmicro.com>