subsys/random: change parameters for timer-based PRNG
PR #36996 disabled running mem_protect/stack_random test on qemu_riscv32
platform because of this test consistently failing on said platform.
This test starts new threads in equal time intervals, and because of
that we get similar results while performing the modulus operation when
calculating stack pointer address.

This can be solved either by adding a delay, changing the number of
iterations or improving the timer-based PRNG by adding multiplier,
increment and modulus parameters.

This commit adds multiplier, increment and modulus parameters to the
timer-based PRNG to address this issue. The values of these parameters
are taken from Pierre L'Ecuyer's work [1] and from an errata [2] to his
article.

Fixes #37006.

[1] Pierre L'Ecuyer, Tables Of Linear Congruential Generators Of
Different Sizes And Good Lattice Structure, 1999,
https://www.ams.org/journals/mcom/1999-68-225/S0025-5718-99-00996-5/S0025-5718-99-00996-5.pdf
[last access: 2021/08/06]
[2]
https://www.iro.umontreal.ca/~lecuyer/myftp/papers/latrules99Errata.pdf
[last access: 2021/08/06]

Signed-off-by: Filip Kokosinski <fkokosinski@antmicro.com>