boards: riscv: kosagi_fomu: Make shell_module work
This is a workaround. Shell init happens in POST_KERNEL with priority 0.
Serial driver Fomu uses is also initialized in POST_KERNEL, but with
priority of 50.

Shell checks whether serial drivers is initialized and otherwise fails.
This commit changes shell init priority to 51, which is one more than
serial driver priority.